natb/pyproject.toml
[project]
name = "natb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi[standard]>=0.116.1",
    "pydantic-settings>=2.10.1",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.42",
    "strawberry-graphql>=0.283.3",
]
